WIP: moved update regions check to MCI
