+++ subsys/suit/mci/CMakeLists.txt
@@ -17,3 +17,4 @@ zephyr_library_sources(src/suit_generic_ids.c)
 zephyr_library_link_libraries(suit_mci)
 zephyr_library_link_libraries_ifdef(CONFIG_SUIT_STORAGE suit_storage_interface)
 zephyr_library_link_libraries(suit_execution_mode)
+zephyr_library_link_libraries(suit_memory_layout_interface)
