+++ subsys/suit/orchestrator/CMakeLists.txt
@@ -20,4 +20,3 @@ zephyr_library_link_libraries(suit_mci)
 zephyr_library_link_libraries(suit_platform_err)
 zephyr_library_link_libraries(suit_execution_mode)
 zephyr_library_link_libraries(suit_cache_interface)
-zephyr_library_link_libraries(suit_memory_layout_interface)
