+++ subsys/suit/orchestrator/src/suit_orchestrator_sdfw.c
@@ -20,8 +20,6 @@
 #include "suit_plat_err.h"
 #include <suit_execution_mode.h>
 #include <suit_dfu_cache.h>
-#include <suit_memory_layout.h>
-#include <sdfw/arbiter.h>
 
 LOG_MODULE_REGISTER(suit_orchestrator, CONFIG_SUIT_LOG_LEVEL);
 
@@ -88,44 +86,6 @@ static void leave_emergency_recovery(void)
 	}
 }
 
-static bool is_address_range_allowed_for_update_region(const uint8_t *address, size_t size)
-{
-	/** The update region (update candidate or DFU cache) address is considered valid in three cases:
-	 * 1) address in external flash
-	 * 2) address accessible by application core
-	 * 3) address accessible by radio core
-	 *  
-	 * Other cases can lead to security issues, leading to potential leaks
-	 * of protected data.
-	 */
-	
-	if (suit_memory_global_address_range_is_in_external_memory((uintptr_t) address, size)) {
-		return true;
-	}
-
-	struct arbiter_mem_params_access mem_params = {
-		.allowed_types = ARBITER_MEM_TYPE(RESERVED, FIXED, FREE),
-		.access = {
-				.owner = NRF_OWNER_APPLICATION,
-				.permissions = ARBITER_MEM_PERM(READ, SECURE),
-				.address = (uintptr_t)address,
-				.size = size,
-			},
-	};
-
-	if (arbiter_mem_access_check(&mem_params) == ARBITER_STATUS_OK) {
-		return true;
-	}
-
-	mem_params.access.owner = NRF_OWNER_RADIOCORE;
-
-	if (arbiter_mem_access_check(&mem_params) == ARBITER_STATUS_OK) {
-		return true;
-	}
-
-	return false;
-}
-
 static int validate_update_candidate_address_and_size(const uint8_t *addr, size_t size)
 {
 	if (addr == NULL || addr == (void *)EMPTY_STORAGE_VALUE) {
@@ -138,7 +98,7 @@ static int validate_update_candidate_address_and_size(const uint8_t *addr, size_
 		return -EFAULT;
 	}
 
-	if (!is_address_range_allowed_for_update_region(addr, size))
+	if (SUIT_PLAT_SUCCESS !=  suit_mci_update_region_address_range_validate(addr, size, false))
 	{
 		LOG_ERR("Update candidate outside of allowed memory range");
 		return -EPERM;
@@ -163,8 +123,8 @@ static int initialize_dfu_cache(const suit_plat_mreg_t *update_regions, size_t u
 
 	for (size_t i = 1; i < update_regions_len; i++) {
 		
-		if (!is_address_range_allowed_for_update_region((uint8_t *)update_regions[i].mem,
-			update_regions[i].size))
+		if (SUIT_PLAT_SUCCESS !=  suit_mci_update_region_address_range_validate(
+			(uint8_t *)update_regions[i].mem, update_regions[i].size, true))
 		{
 			LOG_ERR("Cache pool %d outside of allowed memory range", i);
 			return -EPERM;
